statically link libunwind
Calling dlopen on libunwind doesn't work with the musl libc that we're using
to distribute wheels on linux. Instead statically link in libunwind for musl,
and directly dynamically link against it for gnu libc.

Also add a cfg flag to easily set if we have unwind support, and disable
compiling code if it is set. This should hopefully also fix the ARM build,
and make it easier to port towards freebsd.