Retrieve symbols for binaries missing eh_frame_hdr
On anaconda 3.7.2, the libmkl_vml_avx512.so file doesn't contain an
eh_frame_hdr section. This was causing us to fail to load the binary,
and causing any addresses from there to display as '?'. This also
caused us to continually try to reload the list of binaries even
if not necessary.

Fix by unwind_info optional (falling back to libunwind in this
case).