linux pthreadid reuse fix
On linux, the python threadids are reused after the thread exits. This causes
issues with the cache of python thread -> native thread ids, since the
native thread id is different. With native extension profiling, this causes
errors when we try to get the native stack on a thread that has exitted.
This also will lead to incorrect results with the thread idle detection
code.

Fix and add a integration test to test this out. Also refactor the libunwind
code to return better error messages when this happens.