Don't try to match threads for docker from host
When profiling a process inside a docker container from the Host OS, we fail to
unwind the native stack (which we use for figuring out the native thread id).
This caused an issue with the log being spammed with errors about this obscuring
the actual output (#168).

Fix by falling back to previous behaviour in this case, and also downgrade the
error message to a warning.