Wait for SIGSTOP during attach, reinjecting unrelated signals
We may be reported of another signal(s) first, delivered concurrently to the tracee.
If we don't reinject them with PTRACE_CONT, the tracee won't observe them (thus,
they will be lost).
Furthermore, if we don't consume the SIGSTOP generated by PTRACE_ATTACH,
it will be delivered to the tracee when we PTRACE_DETACH, leaving it stopped.

I followed what's explained in ptrace(2) manpage, under the section "Attaching
and detaching".