Remove syscall lookup code on windows
Instead of trying to lookup syscalls for windows, just assume that code
that is inside a syscall is 'idle' for purposes of reporting in py-spy.

This code was failing on newer versions of windows, causing test breakage in
py-spy. The source files used to generate it hadn't been updated though,